DB 들어가는 숫자 long 에서 int 로 통합

+++ src/main/java/com/ohmycar/domain/BoardVO.java
@@ -6,7 +6,7 @@ import lombok.Data;
 
 @Data
 public class BoardVO {
-	private Long bno; // 게시글 ID
+	private int bno; // 게시글 ID
 	private String title; // 게시글 제목
 	private String content; // 게시글 내용
 	private String writer; // 작성자
